requirements.txt
django~=2.2
django-import-export